+++ levee/core/io.lua
@@ -13,55 +13,6 @@ local MIN_SPLICE_SIZE = 4 * _.pagesize
 
 
 --
--- Iovec
-
-local iovec = ffi.typeof("struct iovec[?]")
-
-
-local Iovec_mt = {}
-Iovec_mt.__index = Iovec_mt
-
-
-function Iovec_mt:write(buf, len)
-	assert(self.n < self.size)
-
-	if not len then
-		len = #buf
-	end
-
-	if len == 0 then
-		return nil, 0
-	end
-
-	if type(buf) == "string" then
-		buf = ffi.cast("char*", buf)
-	end
-
-	self.iov[self.n].iov_base = buf
-	self.iov[self.n].iov_len = len
-
-	self.len = self.len + len
-	self.n = self.n + 1
-end
-
-
-function Iovec_mt:reset()
-	self.n = 0
-	self.len = 0
-end
-
-
-local function Iovec(size)
-	local self = setmetatable({
-		iov = iovec(size),
-		n = 0,
-		len = 0,
-		size = size, }, Iovec_mt)
-	return self
-end
-
-
---
 -- Read
 --
 local R_mt = {}
@@ -305,7 +256,7 @@ function W_mt:iov(size)
 		self.empty = q.empty
 
 		self.hub:spawn(function()
-			local iov = Iovec(size)
+			local iov = d.Iovec(size)
 
 			while true do
 				local err = q:recv()
@@ -313,15 +264,10 @@ function W_mt:iov(size)
 
 				local num = #q
 				for s in q:iter() do
-					-- TODO - eep
-					if s.value then
-						iov:write(s:value())
-					else
-						iov:write(s)
-					end
+					iov:write(s)
 				end
 
-				local err, n = self:writev(iov.iov, iov.n)
+				local err, n = self:writev(iov:value())
 				if err then
 					q.fifo:remove(#q.fifo)   -- this should be handled in stalk
 					self:close()
@@ -958,7 +904,7 @@ function IO_mt:open(name, ...)
 end
 
 
-IO_mt.iovec = Iovec
+IO_mt.iovec = d.Iovec
 IO_mt.R_mt = R_mt