+++ levee/d/init.lua
@@ -1,5 +1,6 @@
 return {
 	Buffer = require("levee.d.buffer"),
+	Iovec = require("levee.d.iovec"),
 	Data = require("levee.d.data"),
 	Fifo = require("levee.d.fifo"),
 	Heap = require("levee.d.heap").Heap,